address review comments to fix API names
