+++ tests/apichecks/testdata/exceptions/acronyms.txt
@@ -389,22 +389,6 @@
 [acronyms] crd=dataprocclusters.dataproc.cnrm.cloud.google.com version=v1beta1: field ".status.config.masterConfig.instanceReferences[].instanceId" should be ".status.config.masterConfig.instanceReferences[].instanceID"
 [acronyms] crd=dataprocclusters.dataproc.cnrm.cloud.google.com version=v1beta1: field ".status.config.secondaryWorkerConfig.instanceReferences[].instanceId" should be ".status.config.secondaryWorkerConfig.instanceReferences[].instanceID"
 [acronyms] crd=dataprocclusters.dataproc.cnrm.cloud.google.com version=v1beta1: field ".status.config.workerConfig.instanceReferences[].instanceId" should be ".status.config.workerConfig.instanceReferences[].instanceID"
-[acronyms] crd=dataprocjobs.dataproc.cnrm.cloud.google.com version=v1alpha1: field ".spec.flinkJob.jarFileUris" should be ".spec.flinkJob.jarFileURIs"
-[acronyms] crd=dataprocjobs.dataproc.cnrm.cloud.google.com version=v1alpha1: field ".spec.hadoopJob.archiveUris" should be ".spec.hadoopJob.archiveURIs"
-[acronyms] crd=dataprocjobs.dataproc.cnrm.cloud.google.com version=v1alpha1: field ".spec.hadoopJob.fileUris" should be ".spec.hadoopJob.fileURIs"
-[acronyms] crd=dataprocjobs.dataproc.cnrm.cloud.google.com version=v1alpha1: field ".spec.hadoopJob.jarFileUris" should be ".spec.hadoopJob.jarFileURIs"
-[acronyms] crd=dataprocjobs.dataproc.cnrm.cloud.google.com version=v1alpha1: field ".spec.hiveJob.jarFileUris" should be ".spec.hiveJob.jarFileURIs"
-[acronyms] crd=dataprocjobs.dataproc.cnrm.cloud.google.com version=v1alpha1: field ".spec.pigJob.jarFileUris" should be ".spec.pigJob.jarFileURIs"
-[acronyms] crd=dataprocjobs.dataproc.cnrm.cloud.google.com version=v1alpha1: field ".spec.pysparkJob.archiveUris" should be ".spec.pysparkJob.archiveURIs"
-[acronyms] crd=dataprocjobs.dataproc.cnrm.cloud.google.com version=v1alpha1: field ".spec.pysparkJob.fileUris" should be ".spec.pysparkJob.fileURIs"
-[acronyms] crd=dataprocjobs.dataproc.cnrm.cloud.google.com version=v1alpha1: field ".spec.pysparkJob.jarFileUris" should be ".spec.pysparkJob.jarFileURIs"
-[acronyms] crd=dataprocjobs.dataproc.cnrm.cloud.google.com version=v1alpha1: field ".spec.pysparkJob.pythonFileUris" should be ".spec.pysparkJob.pythonFileURIs"
-[acronyms] crd=dataprocjobs.dataproc.cnrm.cloud.google.com version=v1alpha1: field ".spec.sparkJob.archiveUris" should be ".spec.sparkJob.archiveURIs"
-[acronyms] crd=dataprocjobs.dataproc.cnrm.cloud.google.com version=v1alpha1: field ".spec.sparkJob.fileUris" should be ".spec.sparkJob.fileURIs"
-[acronyms] crd=dataprocjobs.dataproc.cnrm.cloud.google.com version=v1alpha1: field ".spec.sparkJob.jarFileUris" should be ".spec.sparkJob.jarFileURIs"
-[acronyms] crd=dataprocjobs.dataproc.cnrm.cloud.google.com version=v1alpha1: field ".spec.sparkRJob.archiveUris" should be ".spec.sparkRJob.archiveURIs"
-[acronyms] crd=dataprocjobs.dataproc.cnrm.cloud.google.com version=v1alpha1: field ".spec.sparkRJob.fileUris" should be ".spec.sparkRJob.fileURIs"
-[acronyms] crd=dataprocjobs.dataproc.cnrm.cloud.google.com version=v1alpha1: field ".spec.sparkSQLJob.jarFileUris" should be ".spec.sparkSQLJob.jarFileURIs"
 [acronyms] crd=dataprocworkflowtemplates.dataproc.cnrm.cloud.google.com version=v1beta1: field ".spec.jobs[].hadoopJob.archiveUris" should be ".spec.jobs[].hadoopJob.archiveURIs"
 [acronyms] crd=dataprocworkflowtemplates.dataproc.cnrm.cloud.google.com version=v1beta1: field ".spec.jobs[].hadoopJob.fileUris" should be ".spec.jobs[].hadoopJob.fileURIs"
 [acronyms] crd=dataprocworkflowtemplates.dataproc.cnrm.cloud.google.com version=v1beta1: field ".spec.jobs[].hadoopJob.jarFileUris" should be ".spec.jobs[].hadoopJob.jarFileURIs"
